Move About tile into Settings page
