+++ lib/l10n/app_en_US.arb
@@ -8,17 +8,18 @@
     "playerNoTrack": "No track",
     "playerArranger": "{arranger} (Arranger)",
     "playerComposer": "{composer} (Composer)",
+    "drawerSettingsTile": "Settings",
+    "settingsPageTitle": "Settings",
+    "settingsAppearanceHeader": "Appearance",
     "themeModeHeader": "Theme",
     "themeModeSystem": "System Theme",
     "themeModeLight": "Light Theme",
     "themeModeDark": "Dark Theme",
-    "drawerSettingsTile": "Settings",
-    "drawerAboutTile": "About",
+    "settingsAboutHeader": "About",
+    "aboutTile": "About",
     "aboutDialogLicense": "Licensed under AGPLv3+, developed by MateusRodCosta",
     "aboutDialogAppDescription": "A player for the Vidya Intarweb Playlist (aka VIP Aersia)",
     "aboutDialogVipCats777": "Vidya Intarweb Playlist by Cats777",
     "aboutDialogCopyrightNotice": "All Tracks © & ℗ Their Respective Owners",
-    "aboutDialogSourceCode": "Source code is available at ",
-    "settingsPageTitle": "Settings",
-    "settingsAppearanceHeader": "Appearance"
+    "aboutDialogSourceCode": "Source code is available at "
 }
