+++ lib/l10n/app_pt_BR.arb
@@ -8,17 +8,18 @@
     "playerNoTrack": "Nenhuma faixa",
     "playerArranger": "{arranger} (Arranjador)",
     "playerComposer": "{composer} (Compositor)",
+    "drawerSettingsTile": "Configurações",
+    "settingsPageTitle": "Configurações",
+    "settingsAppearanceHeader": "Aparência",
     "themeModeHeader": "Tema",
     "themeModeSystem": "Tema do Sistema",
     "themeModeLight": "Tema Claro",
     "themeModeDark": "Tema Escuro",
-    "drawerSettingsTile": "Configurações",
-    "drawerAboutTile": "Sobre",
+    "settingsAboutHeader": "Sobre",
+    "aboutTile": "Sobre",
     "aboutDialogLicense": "Licenciado sob a AGPLv3+, desenvolvido por MateusRodCosta",
     "aboutDialogAppDescription": "Um player para a Vidya Intarweb Playlist (aka VIP Aersia)",
     "aboutDialogVipCats777": "Vidya Intarweb Playlist por Cats777",
     "aboutDialogCopyrightNotice": "Todas as Faixas © & ℗ Seus Respectivos Donos",
-    "aboutDialogSourceCode": "O código fonte está disponível em ",
-    "settingsPageTitle": "Configurações",
-    "settingsAppearanceHeader": "Aparência"
+    "aboutDialogSourceCode": "O código fonte está disponível em "
 }
